requirements.txt
autograd==1.5
gym==0.22.0
jax==0.4.12
jaxlib==0.4.12
numpy==1.24.3
Pillow==9.5.0
pygame==2.4.0